% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccLong.R
\name{BCC.multi}
\alias{BCC.multi}
\title{Compute a Bayesian Consensus Clustering model for mixed-type longitudinal data}
\usage{
BCC.multi(
  mydat,
  id,
  time,
  num.cluster,
  formula,
  dist,
  alpha.common,
  initials = NULL,
  sig.var = 0,
  sigma.sq.e.common = 1,
  hyper.par = list(delta = 1, a.star = 1, b.star = 1, aa0 = 0.001, bb0 = 0.001, cc0 =
    0.001, ww0 = 0, vv0 = 1000, dd0 = 0.001, rr0 = 4, RR0 = 3),
  c.ga.tuning = list(1, 1, 1),
  c.theta.tuning = list(1, 1, 1),
  adaptive.tuning = 1,
  align.clusters = 1,
  tuning.freq = 20,
  initial.cluster.membership = "random",
  input.initial.cluster.membership = NULL,
  initial.global.cluster.membership = NULL,
  seed.initial = 2080,
  print.info = "TRUE",
  burn.in,
  thin,
  per,
  max.iter
)
}
\arguments{
\item{mydat}{List of R outcomes (R is the number of outcome, R>=2)}

\item{id}{id-variable: starting from 1 to N}

\item{time}{time variable}

\item{num.cluster}{number of cluster}

\item{formula}{fixed and random effect}

\item{dist}{"gaussian","poisson","binomial", distribution of the outcome}

\item{alpha.common}{1 - common alpha, 0 - separate alphas for each outcome}

\item{initials}{List of initials for: zz, zz.local ga, sigma.sq.u, sigma.sq.e,
Default is NULL}

\item{sig.var}{1 - unstructure random effect variance,
0 - diagonal random effect variance structure,
default is 0}

\item{sigma.sq.e.common}{1 - estimate common residual variance across all groups,
0 - estimate distinct residual variance, default is 1}

\item{hyper.par}{Hyper parameter contains a list of variables includes
delta  = 1, a.star = 1, b.star = 1, aa0  = 1e-3, bb0 = 1e-3,
cc0 = 1e-3, ww0 = 0, vv0 = 1e3, dd0 = 1e-3, rr0 = 4, RR0 = 3}

\item{c.ga.tuning}{tuning parameter for MH algorithm (fixed effect parameters),
each parameter corresponds to an outcome/marker, default
value equals list(1,1,1)}

\item{c.theta.tuning}{tuning parameter for MH algorithm (random effect),
each parameter corresponds to an outcome/marker,
default value equals list(1,1,1)}

\item{adaptive.tuning}{adaptive tuning parameters, 1 - yes, 0 - no,
default is 1}

\item{align.clusters}{assign clusters, default is 1}

\item{tuning.freq}{tuning frequency, default is 20}

\item{initial.cluster.membership}{"mixAK" or "random" or "PAM" or "input" -
input initial cluster membership for local
clustering, default is "random"}

\item{input.initial.cluster.membership}{if use "input",
option input.initial.cluster.membership
must not be empty, default is NULL}

\item{initial.global.cluster.membership}{input initial cluster
membership for global clustering
default is NULL}

\item{seed.initial}{seed for initial clustering
(for initial.cluster.membership = "mixAK")
default is 2080}

\item{print.info}{print model information at each iteration, default is true}

\item{burn.in}{number of samples discarded}

\item{thin}{thinning}

\item{per}{output information every "per" iteration}

\item{max.iter}{maximum number of iteration}
}
\value{
Returns a model contains clustering information
}
\description{
A function that performs clustering on mixed-type (continuous, discrete and
categorical) longitudinal markers using Bayesian consensus clustering method
with MCMC sampling
}
