% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{conRes}
\alias{conRes}
\title{conRes dataset}
\format{
This is a dataframe with two columns and twenty observations
}
\usage{
data(conRes)
}
\description{
This data sets contains the result that run from \code{BayesT} function using epil1 BCC object.
The epil1 object was obtained using \code{BCC.multi} function
}
\examples{
data(conRes)
conRes
}
\keyword{datasets}
