% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{epil}
\alias{epil}
\title{epil dataset}
\format{
This is a dataframe with 4 varaibles and 1852 observations
}
\usage{
data(epil)
}
\description{
This is epileptic.qol data set from \code{joinrRML}
}
\examples{
data(epil)
epil
}
\keyword{datasets}
