% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{epil1}
\alias{epil1}
\title{epil1 model}
\format{
This is a BCC model with thirty elements
}
\usage{
data(epil1)
}
\description{
This model contains the result that run from \code{BCC.multi} function using
epileptic.qol dataset in \code{joinrRML} package.
This model has formula of \code{formula =list(y ~ time + (1|id))}
}
\examples{
data(epil1)
epil1
}
\keyword{datasets}
