% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{epil2}
\alias{epil2}
\title{epil2 model}
\format{
This is a BCC model with thirty elements
}
\usage{
data(epil2)
}
\description{
This model contains the result that run from \code{BCC.multi} function using
epileptic.qol dataset in \code{joinrRML} package.
This model has formula of \code{formula =list(y ~ time + (1 + time|id))}
}
\examples{
data(epil2)
epil2
}
\keyword{datasets}
