% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{epil3}
\alias{epil3}
\title{epil3 model}
\format{
This is a BCC model with thirty elements
}
\usage{
data(epil3)
}
\description{
This model contains the result that run from \code{BCC.multi} function using
epileptic.qol dataset in \code{joinrRML} package.
This model has formula of \code{formula =list(y ~ time + time2 + (1 + time|id))}
}
\examples{
data(epil3)
epil3
}
\keyword{datasets}
