% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{example1}
\alias{example1}
\title{example1 model}
\format{
This is a BCC model with thirty elements
}
\usage{
data(example1)
}
\description{
This is an example model which contains the result that run from \code{BCC.multi}
function using epileptic.qol dataset in \code{joinrRML} package.
Only used the tests. Since small number of iterations
were used, this model can may not represent the true performance
for this method.
}
\examples{
data(example1)
example1
}
\keyword{datasets}
