% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{model.selection.criteria}
\alias{model.selection.criteria}
\title{Model Selection}
\usage{
model.selection.criteria(fit, fast_version = 1)
}
\arguments{
\item{fit}{an objective output from BCC.multi() function}

\item{fast_version}{???}
}
\value{
Returns the calculated score
}
\description{
A function that calculates of DIC and WAIC
}
