% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{PBCseqfit}
\alias{PBCseqfit}
\title{PBCseqfit model}
\format{
This is a BCC model with thirty elements
}
\usage{
data(PBCseqfit)
}
\description{
This model contains the result that run from \code{BCC.multi} function using
PBC910 dataset in \code{mixAK} package
}
\examples{
data(PBCseqfit)
PBCseqfit
}
\keyword{datasets}
