% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{plot.BCC}
\alias{plot.BCC}
\title{Generic plot method for BCC objects}
\usage{
\method{plot}{BCC}(x, ...)
}
\arguments{
\item{x}{An object of class BCC.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Void function plot model object, no object return
}
\description{
Generic plot method for BCC objects
}
\examples{
# get data from the package
data(epil1)
fit.BCC <- epil1
plot(fit.BCC)
}
