% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{print.BCC}
\alias{print.BCC}
\title{Generic print method for BCC objects}
\usage{
\method{print}{BCC}(x, ...)
}
\arguments{
\item{x}{An object of class BCC.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Void function prints model information, no object return
}
\description{
Generic print method for BCC objects
}
\examples{
# get data from the package
data(epil2)
fit.BCC <- epil2
print(fit.BCC)
}
