% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethods.R
\name{summary.BCC}
\alias{summary.BCC}
\title{Generic summary method for BCC objects}
\usage{
\method{summary}{BCC}(object, ...)
}
\arguments{
\item{object}{An object of class BCC.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Void function summarize model information, no object return
}
\description{
Generic summary method for BCC objects
}
\examples{
# get data from the package
data(epil2)
fit.BCC <- epil2
summary(fit.BCC)
}
